pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai==1.12.1",
    "pandas>=2.3.2",
    "python-dotenv==1.1.0",
]
